# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go Clean Architecture API with Better Auth integration",
        "title": "Next-Go-PG API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/webhooks/send-2fa-enabled": {
            "post": {
                "description": "Notifies user that 2FA has been enabled on their account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Send 2FA enabled notification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook secret for authentication",
                        "name": "X-Webhook-Secret",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Notification data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.Send2FAEnabledNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhooks/send-2fa-otp": {
            "post": {
                "description": "Called by Better Auth to send 2FA one-time password via email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Send 2FA OTP email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook secret for authentication",
                        "name": "X-Webhook-Secret",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "OTP data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.Send2FAOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhooks/send-magic-link": {
            "post": {
                "description": "Called by Better Auth to send magic link login email",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Send magic link email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook secret for authentication",
                        "name": "X-Webhook-Secret",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Magic link data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.SendMagicLinkRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhooks/send-passkey-added": {
            "post": {
                "description": "Notifies user that a new passkey has been added to their account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Send passkey added notification",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook secret for authentication",
                        "name": "X-Webhook-Secret",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Notification data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.SendPasskeyAddedNotificationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhooks/send-verification-email": {
            "post": {
                "description": "Called by Better Auth to send email verification link",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Send verification email",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook secret for authentication",
                        "name": "X-Webhook-Secret",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Verification data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.SendVerificationEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/webhooks/session-created": {
            "post": {
                "description": "Called by Better Auth when a new session is created. Sends notification email for new devices.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Handle new session webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Webhook secret for authentication",
                        "name": "X-Webhook-Secret",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Session data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.SessionCreatedRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.MessageResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/notifications_interfaces_http.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "notifications_interfaces_http.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "unauthorized"
                }
            }
        },
        "notifications_interfaces_http.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Hello World"
                }
            }
        },
        "notifications_interfaces_http.Send2FAEnabledNotificationRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "method": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "notifications_interfaces_http.Send2FAOTPRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "otp": {
                    "type": "string"
                }
            }
        },
        "notifications_interfaces_http.SendMagicLinkRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "notifications_interfaces_http.SendPasskeyAddedNotificationRequest": {
            "type": "object",
            "properties": {
                "device": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "passkeyName": {
                    "type": "string"
                }
            }
        },
        "notifications_interfaces_http.SendVerificationEmailRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "notifications_interfaces_http.SessionCreatedRequest": {
            "type": "object",
            "properties": {
                "ipAddress": {
                    "type": "string"
                },
                "sessionId": {
                    "type": "string"
                },
                "userAgent": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}